amplify/custom/agents/newAgent/src/model/load.py
from strands.models import BedrockModel

# Uses Amazon Nova Micro via cross-region inference profile
# https://docs.aws.amazon.com/bedrock/latest/userguide/inference-profiles-support.html
MODEL_ID = "global.anthropic.claude-haiku-4-5-20251001-v1:0"

def load_model() -> BedrockModel:
    """
    Get Bedrock model client.
    Uses IAM authentication via the execution role.
    """
    return BedrockModel(model_id=MODEL_ID)
    #